fix: BIGINT

+++ golang/cmd/kafka-to-postgresql-v2/postgresql/analytics-work-order.go
@@ -42,11 +42,11 @@ func (c *Connection) InsertWorkOrderCreate(msg *sharedStructs.WorkOrderCreateMes
 					 $4,
 					 $5,
 					 CASE
-					   WHEN $6 :: INT IS NOT NULL THEN to_timestamp($6 :: INT / 1000)
+					   WHEN $6 :: BIGINT IS NOT NULL THEN to_timestamp($6 :: BIGINT / 1000)
 					   ELSE NULL
 					 END :: timestamptz,
 					 CASE
-					   WHEN $7 :: INT IS NOT NULL THEN to_timestamp($7 :: INT / 1000)
+					   WHEN $7 :: BIGINT IS NOT NULL THEN to_timestamp($7 :: BIGINT / 1000)
 					   ELSE NULL
 					 END :: timestamptz) 
 	`, values...)
